fix: use uint64 bounds arithmetic to prevent 32-bit offset-truncation OOB (#74)
On 32-bit platforms (GOARCH=386/arm), int(uint32 >= 2^31) is negative.
The pattern `if int(off)+k > len(b)` silently passes for such offsets,
then the subsequent slice with the full uint32 panics (OOB).

Affected locations:
- exif/ifd.go parseSingleIFD, traverse: change int(offset)+2 > len(b)
  to uint64(offset)+2 > uint64(len(b))
- format/detect.go parseTIFFScanHeader: same pattern for ifd0Off
- format/png/png.go readChunk: check rawLen > math.MaxInt32 before
  int cast; return ErrChunkTooLarge (PNG ISO 15948 §11.2.1 caps at 2^31-1)
- format/webp/webp.go collectOriginalChunks: same guard for RIFF chunk
  size; break early when rawSize > math.MaxInt32

All fixes mirror the pattern already used in exif/ifd.go parseIFDEntry
(lines 91-95) and the IPTC IRB parser (task #45).

Regression gate (task #74 acceptance criteria):
- TestParseHighOffsetIFD (exif): 5 subtests with offsets 0x80000000/0xFFFFFFFF
- TestDetectTIFFHighIFD0Offset (format): 4 subtests; verifies FormatTIFF fallback
- TestPNGReadChunkLargeLength (format/png): 4 subtests; asserts ErrChunkTooLarge
- TestCollectOriginalChunksLargeSize (format/webp): 3 subtests; no panic
Fuzz seeds added to FuzzParseEXIF, FuzzTIFFExtract, FuzzPNGExtract, FuzzWebPExtract.

GOARCH=386 GOOS=linux go build ./...: clean
go test -race ./...: all 36 packages pass
golangci-lint (clean cache): 0 issues
staticcheck: 0 issues

+++ exif/ifd.go
@@ -122,7 +122,11 @@ func parseIFDEntry(b []byte, e int, order binary.ByteOrder) (IFDEntry, bool) {
 //
 // Callers are responsible for cycle detection before calling this function.
 func parseSingleIFD(b []byte, offset uint32, order binary.ByteOrder) (*IFD, uint32, bool) {
-	if int(offset)+2 > len(b) {
+	// Use uint64 arithmetic to avoid int truncation on 32-bit platforms
+	// (GOARCH=386/arm): int(uint32 >= 2^31) is negative, which would cause the
+	// guard to pass while the subsequent slice panics. Performing the comparison
+	// in uint64 is safe on all platforms (task #74).
+	if uint64(offset)+2 > uint64(len(b)) {
 		return nil, 0, false
 	}
 
@@ -197,7 +201,11 @@ func extractJPEGThumbnail(b []byte, ifd *IFD, order binary.ByteOrder) []byte {
 // The next-IFD pointer chain is followed iteratively (not recursively) to
 // prevent stack overflows on cyclic or deeply nested inputs (fuzz safety).
 func traverse(b []byte, offset uint32, order binary.ByteOrder) (*IFD, error) {
-	if int(offset)+2 > len(b) {
+	// Use uint64 arithmetic to avoid int truncation on 32-bit platforms
+	// (GOARCH=386/arm): int(uint32 >= 2^31) is negative, which would cause the
+	// guard to pass while the subsequent slice panics. Performing the comparison
+	// in uint64 is safe on all platforms (task #74).
+	if uint64(offset)+2 > uint64(len(b)) {
 		return nil, &metaerr.CorruptMetadataError{
 			Format: "EXIF",
 			Reason: fmt.Sprintf("IFD offset %d out of bounds (buf len %d)", offset, len(b)),
